remove duplicate components
