--- src/components/app/BurgerIngredients/BunsSelect/Bun/Bun.module.css
@@ -1,19 +1,0 @@
-.ingredient {
-    width: 272px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.price {
-margin: 4px auto 0;
-font-size:25px;
-}
-
-.name {
-    font-size:16px;
-    line-height: 24px;
-    font-weight: 400;
-}
