--- src/components/app/BurgerIngredients/Main/Filling/Filling.module.css
@@ -1,21 +1,0 @@
-.ingredient {
-    width: 250px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.price {
-    margin: 4px auto 0;
-    font-size: 25px;
-}
-
-.name {
-    font-size: 16px;
-    line-height: 24px;
-    font-weight: 400;
-    text-wrap: wrap;
-    text-align: center;
-}
