--- src/components/app/BurgerIngredients/Main/Main.module.css
@@ -1,37 +1,0 @@
-.title {
-    font-size: 24px;
-    font-weight: 700;
-    line-height: 30px;
-}
-
-.container {
-    display: grid;
-    grid-template-columns: repeat(2, 250px);
-    column-gap: 24px;
-    row-gap: 32px;
-    align-items: center;
-    justify-content: center;
-    height: 600px;
-}
-
-.ingredient {
-    width: 250px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.price {
-    margin: 4px auto 0;
-    font-size: 25px;
-}
-
-.name {
-    font-size: 16px;
-    line-height: 24px;
-    font-weight: 400;
-    text-wrap: wrap;
-    text-align: center;
-}
