--- src/components/app/BurgerIngredients/Souses/Souses.module.css
@@ -1,15 +1,0 @@
-.title {
-   font-size: 24px;
-   font-weight: 700;
-   line-height: 30px; 
-}
-
-.container {
-    display: grid;
-    grid-template-columns: repeat(2, 250px);
-    column-gap: 24px;
-    row-gap: 32px;
-    align-items: center;
-    justify-content: center;
-    height: 600px;
-}
